Edits

+++ Slurm_Submission/QUBO/Modules/Energy_Encoder_Classes.py
@@ -138,7 +138,7 @@ class BVAE(pl.LightningModule):
         shifted_vector = transitioned_vectors + self.shift
 
         """"""
-        transitioned_vectors_with_gradient = self.scale_vector_copy_gradient(shifted_vector, probabilities)
+        transitioned_vectors_with_gradient = self.scale_vector_copy_gradient(shifted_vector.type(torch.int8), probabilities)
         """"""
 
         x_hat = self.vae.decode(transitioned_vectors_with_gradient)
